# Conflicts:
#	app/src/main/java/com/rain/sdk/sample/NetworkClient.kt
#	app/src/main/java/com/rain/sdk/sample/SampleViewModel.kt
#	app/src/main/java/com/rain/sdk/sample/screens/BalancesViewModel.kt
#	app/src/main/java/com/rain/sdk/sample/screens/CollateralWithdrawViewModel.kt
#	app/src/main/java/com/rain/sdk/sample/screens/SendTokensViewModel.kt
#	docs/METHODS.md
#	rain-sdk/src/main/java/com/rain/sdk/interfaces/RainClient.kt
#	rain-sdk/src/main/java/com/rain/sdk/internal/core/PortalManager.kt
#	rain-sdk/src/main/java/com/rain/sdk/internal/core/RainSdkManager.kt
#	rain-sdk/src/main/java/com/rain/sdk/internal/provider/PortalWalletProvider.kt
#	rain-sdk/src/main/java/com/rain/sdk/internal/provider/WalletProvider.kt
#	rain-sdk/src/main/java/com/rain/sdk/internal/transaction/TransactionCoordinator.kt
#	rain-sdk/src/main/java/com/rain/sdk/utils/EthereumConverter.kt

+++ rain-sdk/src/main/java/com/rain/sdk/internal/provider/PortalWalletProvider.kt
@@ -32,7 +32,7 @@ internal class PortalWalletProvider(
   override suspend fun sendNativeToken(
     chainId: Int,
     toAddress: String,
-    amountInEth: Double
+    amountInEth: BigDecimal
   ): String {
     val fromAddress = getWalletAddress()
     val valueWeiHex = EthereumConverter.convertEthToWeiHex(amountInEth)
@@ -51,13 +51,13 @@ internal class PortalWalletProvider(
     chainId: Int,
     contractAddress: String,
     toAddress: String,
-    amount: Double,
+    amount: BigDecimal,
     decimals: Int
   ): String {
     val fromAddress = getWalletAddress()
 
     // Encode ERC-20 transfer(address, uint256) function call
-    val tokenAmount = amount.toBigDecimal()
+    val tokenAmount = amount
       .multiply(BigDecimal.TEN.pow(decimals))
       .toBigInteger()
     val function = Function(
